add a crude fs barrier to the debug helper
I should not have to do this... :/
