+++ src/radical/utils/misc.py
@@ -4,7 +4,6 @@ import time
 import regex
 import signal
 import threading
-import traceback
 import url as ruu
 
 # ------------------------------------------------------------------------------
@@ -206,120 +205,6 @@ def time_diff (dt_abs, dt_stamp) :
     # get seconds as float 
     seconds = delta.seconds + delta.microseconds/1E6
     return seconds
-
-
-# --------------------------------------------------------------------
-#
-def get_trace():
-
-    trace = sys.exc_info ()[2]
-
-    if  trace :
-        stack           = traceback.extract_tb  (trace)
-        traceback_list  = traceback.format_list (stack)
-        return "".join (traceback_list)
-
-    else :
-        stack           = traceback.extract_stack ()
-        traceback_list  = traceback.format_list (stack)
-        return "".join (traceback_list[:-1])
-
-
-# ------------------------------------------------------------------------------
-#
-class DebugHelper (object) :
-    """
-    When instantiated, and when "RADICAL_DEBUG" is set in the environment, this
-    class will install a signal handler for SIGUSR1.  When that signal is
-    received, a stacktrace for all threads is printed to stdout.
-    Additionally, we check if SIGINFO is available, which is generally bound to CTRL-T.
-    """
-
-    def __init__ (self) :
-
-        if 'MainThread' not in threading.current_thread().name:
-            # python only supports signals in main threads :/
-            return
-
-        if 'RADICAL_DEBUG' in os.environ :
-            signal.signal(signal.SIGUSR1, print_stacktraces) # signum 30
-            signal.signal(signal.SIGQUIT, print_stacktraces) # signum  3
-
-            try:
-                assert signal.SIGINFO
-                signal.signal(signal.SIGINFO, print_stacktraces) # signum 29
-            except AttributeError as e:
-                pass
-
-
-# ------------------------------------------------------------------------------
-#
-def print_stacktraces (signum=None, sigframe=None) :
-    """
-    signum, sigframe exist to satisfy signal handler signature requirements
-    """
-
-    this_tid = threading.currentThread().ident
-
-    # if multiple processes (ie. a process group) get the signal, then all
-    # traces are mixed together.  Thus we waid 'pid%100' milliseconds, in
-    # the hope that this will stagger the prints.
-    pid = int(os.getpid())
-    time.sleep((pid%100)/1000)
-
-    out  = "===============================================================\n"
-    out += "RADICAL Utils -- Debug Helper -- Stacktraces\n"
-    try :
-        info = get_stacktraces ()
-    except Exception as e:
-        out += 'skipping frame'
-        info = None
-
-    if info:
-        for tid, tname in info :
-
-            if tid == this_tid : marker = '[active]'
-            else               : marker = ''
-            out += "---------------------------------------------------------------\n"
-            out += "Thread: %s %s\n" % (tname, marker)
-            out += "  PID : %s \n"   % os.getpid()
-            out += "  TID : %s \n"   % tid
-            for fname, lineno, method, code in info[tid,tname] :
-
-                if code: code = code.strip()
-                else   : code = '<no code>'
-
-              # # [:-1]: .py vs. .pyc :/
-              # if not (__file__[:-1] in fname and \
-              #         method in ['get_stacktraces', 'print_stacktraces']):
-                if method not in ['get_stacktraces', 'print_stacktraces']:
-                    out += "  File: %s, line %d, in %s\n" % (fname, lineno, method)
-                    out += "        %s\n" % code
-
-    out += "==============================================================="
-
-    sys.stdout.write("%s\n" % out)
-
-    if 'RADICAL_DEBUG' in os.environ:
-        with open('/tmp/ru.stacktrace.%s.log' % pid, 'w') as f:
-            f.write ("%s\n" % out)
-
-    return True
-
-
-# --------------------------------------------------------------------------
-#
-def get_stacktraces () :
-
-    id2name = {}
-    for th in threading.enumerate():
-        id2name[th.ident] = th.name
-
-    ret = dict()
-    for tid, stack in sys._current_frames().items():
-        ret[tid,id2name[tid]] = traceback.extract_stack(stack)
-
-    return ret
 
 
 # ------------------------------------------------------------------------------
